+++ src/modules/files/sendfile.c
@@ -246,12 +246,12 @@ static void call_callback_and_free(struct callback *cb, void *this_, void *arg)
  */
 
 /* writev() without the IOV_MAX limit. */
-static int send_iov(int fd, struct iovec *iov, int iovcnt)
+static ptrdiff_t send_iov(int fd, struct iovec *iov, int iovcnt)
 {
-  int sent = 0;
+  ptrdiff_t sent = 0;
 
   while (iovcnt) {
-    int bytes;
+    ptrdiff_t bytes;
     int cnt = iovcnt;
 
 #ifdef IOV_MAX
